.github/workflows/compile-examples-os.yml: Added compile example as per OS.
The workflow runs on the operating system passed as input and compile all sketches for the passed matrix of fqbns (boards) in different jobs. The workflow needs to checkout the Arduino asset (core or library) repository and the arduino-devops repository. The arduino-devops utilities will take care of installings all the necessary dependencies, install arduino-cli (for any operating system) and install the required Arduino cores, before compiling the fqbn-sketch matrix.

Signed-off-by: jaenrig-ifx <enriquezgarcia.external@infineon.com>